Moved component to component folder
