+++ src/lib/components/NumberInput.svelte
