+++ src/lib/components/TextInput.svelte
@@ -4,18 +4,22 @@
   export let placeholder = "";
   export let value: string | null = "";
   export let disabled = false;
+  export let inputHandler = () => {};
 </script>
 
 <div>
-  <label for={id} class="block text-sm font-medium leading-6 text-gray-900"
-    >{label}</label
-  >
+  {#if label !== ""}
+    <label for={id} class="block text-sm font-medium leading-6 text-gray-900"
+      >{label}</label
+    >
+  {/if}
   <div class="mt-2">
     <input
       {id}
       {disabled}
       type="text"
       {placeholder}
+      on:input={inputHandler}
       class="block w-full pl-2 rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6 disabled:bg-gray-100 disabled:text-gray-400"
       bind:value
     />
